bits and bobs

+++ src/main/java/com/jesz/createdieselgenerators/other/SpoutCanisterFilling.java
@@ -4,17 +4,16 @@ import com.jesz.createdieselgenerators.blocks.entity.CanisterBlockEntity;
 import com.jesz.createdieselgenerators.config.ConfigRegistry;
 import com.simibubi.create.api.behaviour.BlockSpoutingBehaviour;
 import com.simibubi.create.content.fluids.spout.SpoutBlockEntity;
+import io.github.fabricators_of_create.porting_lib.util.FluidStack;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.entity.BlockEntity;
-import net.minecraftforge.common.capabilities.ForgeCapabilities;
-import net.minecraftforge.fluids.FluidStack;
-import net.minecraftforge.fluids.capability.IFluidHandler;
 
 public class SpoutCanisterFilling extends BlockSpoutingBehaviour {
+
     @Override
-    public int fillBlock(Level level, BlockPos pos, SpoutBlockEntity spout, FluidStack availableFluid, boolean simulate) {
+    public long fillBlock(Level level, BlockPos pos, SpoutBlockEntity spout, FluidStack availableFluid, boolean simulate) {
         if(!ConfigRegistry.CANISTER_SPOUT_FILLING.get())
             return 0;
         BlockEntity blockEntity = level.getBlockEntity(pos);
